+++ src/rpc/mod.rs
@@ -82,6 +82,8 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::eth::EthGetBlockTransactionCountByNumber);
         $callback!(crate::rpc::eth::EthGetMessageCidByTransactionHash);
         $callback!(crate::rpc::eth::EthGetTransactionCount);
+        $callback!(crate::rpc::eth::EthMaxPriorityFeePerGas);
+        $callback!(crate::rpc::eth::EthProtocolVersion);
         $callback!(crate::rpc::eth::EthGetTransactionHashByCid);
 
         // gas vertical
@@ -169,6 +171,7 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::state::StateSearchMsg);
         $callback!(crate::rpc::state::StateSearchMsgLimited);
         $callback!(crate::rpc::state::StateFetchRoot);
+        $callback!(crate::rpc::state::StateCompute);
         $callback!(crate::rpc::state::StateMinerPreCommitDepositForPower);
         $callback!(crate::rpc::state::StateVerifierStatus);
         $callback!(crate::rpc::state::StateGetClaim);
@@ -196,6 +199,7 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::wallet::WalletNew);
         $callback!(crate::rpc::wallet::WalletSetDefault);
         $callback!(crate::rpc::wallet::WalletSign);
+        $callback!(crate::rpc::wallet::WalletSignMessage);
         $callback!(crate::rpc::wallet::WalletValidateAddress);
         $callback!(crate::rpc::wallet::WalletVerify);
         $callback!(crate::rpc::wallet::WalletDelete);